+++ Data_Analysis/LammpsNeighborList.m
@@ -17,8 +17,10 @@ coord_scl               =   LammpsSclCoord(varargin{1});
 
 %% Constructing Neighbor List
 
-neighbor_list           =   zeros(varargin{1}.num_atoms,varargin{1}.num_atoms,varargin{1}.num_steps_sim);
-neighbor_num            =   zeros(varargin{1}.num_atoms,varargin{1}.num_steps_sim);
+num_tolerance           =   30;
+num_neighbor_max        =   round((4*pi*varargin{2}^3 /3)/(varargin{1}.box_volume) * varargin{1}.num_atoms) + num_tolerance;
+neighbor_list           =   zeros(varargin{1}.num_atoms,num_neighbor_max,varargin{1}.num_steps_sim);
+num_neighbor            =   zeros(varargin{1}.num_atoms,varargin{1}.num_steps_sim);
 
 for atom_01 = 1 : varargin{1}.num_atoms - 1
     for atom_02 =  atom_01 + 1 : varargin{1}.num_atoms
@@ -26,11 +28,11 @@ for atom_01 = 1 : varargin{1}.num_atoms - 1
         r_scl = r_scl - round(r_scl);
         r_diff = sqrt(varargin{1}.box_size' .^2 * r_scl.^2);
     	pos = find(r_diff <= varargin{2});
-        neighbor_num(atom_01,pos) = neighbor_num(atom_01,pos) + 1;
-        neighbor_num(atom_02,pos) = neighbor_num(atom_02,pos) + 1;
+        num_neighbor(atom_01,pos) = num_neighbor(atom_01,pos) + 1;
+        num_neighbor(atom_02,pos) = num_neighbor(atom_02,pos) + 1;
         for i = 1 : length(pos) 
-            neighbor_list(atom_01,neighbor_num(atom_01,pos(i)),pos(i)) = atom_02;
-            neighbor_list(atom_02,neighbor_num(atom_02,pos(i)),pos(i)) = atom_01;
+            neighbor_list(atom_01,num_neighbor(atom_01,pos(i)),pos(i)) = atom_02;
+            neighbor_list(atom_02,num_neighbor(atom_02,pos(i)),pos(i)) = atom_01;
         end
     end
 end
@@ -38,4 +40,4 @@ end
 %% -----------------------Output-----------------------
 
 varargout{1}.neighbor_list      =   neighbor_list;
-varargout{1}.neighbor_num       =   neighbor_num;
+varargout{1}.neighbor_num       =   num_neighbor;