+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 
 - LammpsNeighborList
 - LammpsSclCoord
+- LammpsReadDump will return 'box_volume' in output data struture
+
+### Fixed
+
+- LammpsRDF use data.box_volume instead of 'data.box_size' * data.box_size'.
 
 ## [0.0.3] -11-27-2019